refactor(ui): move combobox to ui package
