+++ packages/ui/src/components/combobox.tsx
@@ -1,8 +1,12 @@
 'use client';
 
-import * as React from 'react';
+import {
+  type FC,
+  type PropsWithChildren,
+  createContext,
+  useContext,
+} from 'react';
 import {Check, ChevronsUpDown} from 'lucide-react';
-
 import {cn} from '../lib/utils';
 import {Button} from './button';
 import {
@@ -14,77 +18,171 @@ import {
   CommandList,
 } from './command';
 import {Popover, PopoverContent, PopoverTrigger} from './popover';
+import {useCombobox, type UseComboboxOptions} from '../hooks/useCombobox';
 
-const frameworks = [
-  {
-    value: 'next.js',
-    label: 'Next.js',
-  },
-  {
-    value: 'sveltekit',
-    label: 'SvelteKit',
-  },
-  {
-    value: 'nuxt.js',
-    label: 'Nuxt.js',
-  },
-  {
-    value: 'remix',
-    label: 'Remix',
-  },
-  {
-    value: 'astro',
-    label: 'Astro',
-  },
-];
+interface ComboboxContextValue {
+  open: boolean;
+  currentValue: string;
+  handleSelect: (value: string) => void;
+}
 
-export function ComboboxDemo() {
-  const [open, setOpen] = React.useState(false);
-  const [value, setValue] = React.useState('');
+const ComboboxContext = createContext<ComboboxContextValue | null>(null);
+
+function useComboboxContext() {
+  const context = useContext(ComboboxContext);
+  if (!context) {
+    throw new Error(
+      'Combobox compound components must be used within Combobox',
+    );
+  }
+  return context;
+}
+
+/**
+ * Root props for the compound Combobox component.
+ */
+export type ComboboxRootProps<T extends string = string> = PropsWithChildren<
+  UseComboboxOptions<T> & {
+    /**
+     * Optional selected value used for item highlighting when it differs from
+     * the controlled value.
+     */
+    currentValue?: string;
+  }
+>;
+
+function ComboboxRoot<T extends string = string>({
+  value,
+  onChange,
+  currentValue,
+  children,
+}: ComboboxRootProps<T>) {
+  const {popoverProps, handleSelect, open} = useCombobox({value, onChange});
 
   return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger asChild>
-        <Button
-          variant="outline"
-          role="combobox"
-          aria-expanded={open}
-          className="w-[200px] justify-between"
-        >
-          {value
-            ? frameworks.find((framework) => framework.value === value)?.label
-            : 'Select framework...'}
-          <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-        </Button>
-      </PopoverTrigger>
-      <PopoverContent className="w-[200px] p-0">
-        <Command>
-          <CommandInput placeholder="Search framework..." />
-          <CommandList>
-            <CommandEmpty>No framework found.</CommandEmpty>
-            <CommandGroup>
-              {frameworks.map((framework) => (
-                <CommandItem
-                  key={framework.value}
-                  value={framework.value}
-                  onSelect={(currentValue) => {
-                    setValue(currentValue === value ? '' : currentValue);
-                    setOpen(false);
-                  }}
-                >
-                  <Check
-                    className={cn(
-                      'mr-2 h-4 w-4',
-                      value === framework.value ? 'opacity-100' : 'opacity-0',
-                    )}
-                  />
-                  {framework.label}
-                </CommandItem>
-              ))}
-            </CommandGroup>
-          </CommandList>
-        </Command>
-      </PopoverContent>
-    </Popover>
+    <ComboboxContext.Provider
+      value={{
+        open,
+        currentValue: currentValue ?? String(value),
+        handleSelect,
+      }}
+    >
+      <Popover {...popoverProps}>{children}</Popover>
+    </ComboboxContext.Provider>
   );
 }
+
+/**
+ * Props for the button that opens the combobox popover.
+ */
+export interface ComboboxTriggerProps {
+  className?: string;
+  ariaLabel?: string;
+}
+
+const ComboboxTrigger: FC<PropsWithChildren<ComboboxTriggerProps>> = ({
+  children,
+  className,
+  ariaLabel,
+}) => {
+  const {open} = useComboboxContext();
+
+  return (
+    <PopoverTrigger asChild>
+      <Button
+        variant="outline"
+        role="combobox"
+        aria-expanded={open}
+        aria-label={ariaLabel}
+        className={cn('h-8 justify-between text-xs font-normal', className)}
+      >
+        {children}
+        <ChevronsUpDown className="ml-2 h-3.5 w-3.5 shrink-0 opacity-50" />
+      </Button>
+    </PopoverTrigger>
+  );
+};
+
+/**
+ * Props for the popover content that contains combobox options.
+ */
+export interface ComboboxContentProps {
+  searchable?: boolean;
+  searchPlaceholder?: string;
+  emptyMessage?: string;
+}
+
+const ComboboxContent: FC<PropsWithChildren<ComboboxContentProps>> = ({
+  children,
+  searchable = false,
+  searchPlaceholder = 'Search...',
+  emptyMessage = 'No matching option.',
+}) => {
+  return (
+    <PopoverContent className="w-[--radix-popover-trigger-width] p-0 text-xs">
+      <Command>
+        {searchable && (
+          <CommandInput placeholder={searchPlaceholder} className="text-xs" />
+        )}
+        <CommandList>
+          <CommandEmpty>{emptyMessage}</CommandEmpty>
+          <CommandGroup>{children}</CommandGroup>
+        </CommandList>
+      </Command>
+    </PopoverContent>
+  );
+};
+
+/**
+ * Props for an individual selectable combobox item.
+ */
+export interface ComboboxItemProps {
+  value: string;
+  isSelected?: boolean;
+}
+
+const ComboboxItem: FC<PropsWithChildren<ComboboxItemProps>> = ({
+  value,
+  children,
+  isSelected,
+}) => {
+  const {currentValue, handleSelect} = useComboboxContext();
+  const selected = isSelected ?? currentValue === value;
+
+  return (
+    <CommandItem value={value} onSelect={handleSelect}>
+      <Check
+        className={cn(
+          'h-3.5 w-3.5 shrink-0',
+          selected ? 'opacity-100' : 'opacity-0',
+        )}
+      />
+      {children}
+    </CommandItem>
+  );
+};
+
+/**
+ * Compound component for Popover-based combobox selectors.
+ *
+ * @example
+ * ```tsx
+ * <Combobox value={value} onChange={setValue}>
+ *   <Combobox.Trigger>
+ *     <span>{selectedLabel}</span>
+ *   </Combobox.Trigger>
+ *   <Combobox.Content searchable searchPlaceholder="Search...">
+ *     {options.map((option) => (
+ *       <Combobox.Item key={option.value} value={option.value}>
+ *         <span>{option.label}</span>
+ *       </Combobox.Item>
+ *     ))}
+ *   </Combobox.Content>
+ * </Combobox>
+ * ```
+ */
+export const Combobox = Object.assign(ComboboxRoot, {
+  Trigger: ComboboxTrigger,
+  Content: ComboboxContent,
+  Item: ComboboxItem,
+});
